rust-paging: refactor the test framework
Split out the test cases to be run inside vm into a dedicated crate
tests/rust-test-paging, which brings several benefits:
1) we could build rust-paging with stable toolchain.
2) we could run tests both on host or inside vm.
3) it's much faster to build the rust-paging crate.
