+++ Cargo.toml
@@ -15,6 +15,7 @@ members = [
     "rust-tdpayload-signing",
     "rust-tdshim-key-enroll",
     "rust-test-payload",
+    "tests/rust-test-paging",
 ]
 
 # the profile used for `cargo build`
