+++ rust-paging/Cargo.toml
@@ -13,10 +13,4 @@ x86 = "0.41.0"
 x86_64 = "=0.14.6"
 log = "0.4.13"
 rust-td-layout = { path = "../rust-td-layout" }
-test_lib = { path = "../test_lib" }
 spin = "0.9.2"
-bootloader = "=0.10.9"
-
-[package.metadata.bootloader]
-map-physical-memory = true
-map-page-table-recursively = true
