+++ tests/rust-test-paging/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "rust-test-paging"
+version = "0.1.0"
+description = "Run rust-paging test cases inside a VM"
+repository = "https://github.com/confidential-containers/td-shim"
+homepage = "https://github.com/confidential-containers"
+edition = "2018"
+license = "BSD-2-Clause-Patent"
+
+[dev-dependencies]
+bootloader = "=0.10.9"
+paging = { path = "../../rust-paging" }
+rust-td-layout = { path = "../../rust-td-layout" }
+test_lib = { path = "../../test_lib" }
+
+[package.metadata.bootloader]
+map-physical-memory = true
+map-page-table-recursively = true
