+++ tests/rust-test-paging/x86_64-custom.json
@@ -1,0 +1,1 @@
+../../devtools/rustc-targets/x86_64-custom.json
